+++ preload.js
@@ -7,6 +7,7 @@ contextBridge.exposeInMainWorld("sns", {
   openTimePage: () => ipcRenderer.invoke("open-time-page"),
   setExclude: (apps) => ipcRenderer.invoke("set-exclude", apps),
   hidePopup: () => ipcRenderer.invoke("hide-popup"),
+  togglePin: () => ipcRenderer.invoke("toggle-pin"),
   dragStart: () => ipcRenderer.invoke("drag-start"),
   dragEnd: () => ipcRenderer.invoke("drag-end"),
   quit: () => ipcRenderer.invoke("quit"),